Add custom img and view to test dispatch
Doesn't work: returning RAI from Interval view doesn't seem to be the
right thing.
